RHMWRT-62 and RHOAR-112: Remove ISTIO examples
Missions and Boosters have been renamed to example applications and some of the functionality of the missions and boosters are now provided as capabilities. Other missions and boosters are also not really good as example applications since they are only focusing on showing a particular feature of Istio/Service Mesh.
The following mission should be removed since these are better surfaced as Istio examples and not connected to a particular runtime.

- Istio - Routing
- Istio - Security
- Istio - Circuit Breaker